final amends
:questioned: Each field includes the necessary attributes. ID, type, and name for inputs. Select/textarea elements have id and name attributes.
I would add an ID and name to the submit button as that's technically an input!
:questioned: Zip code width should be 50% width of the other input fields in mobile view and 25% in tablet view.
Close! On mobile it's a little bit wider than 50%, and on tablet/desktop it looks to be around 50%. Be careful not to just use the 50% and 25% values directly in the width field. The goal here is to make them visually look 1/4 or 1/2 of the other input fields.
:questioned: Roughly matches layout of the mobile and desktop mockups. General spacing and arrangement of the elements matches the design of the mockups for mobile and desktop
It looks like the State field doesn't read the full width as the others in tablet/desktop
:questioned: Uses a Google Font.
Great! I can see that you've imported and used Playfair Display. Just to keep our code super clean, be sure not to import Robot unless we're using it!